Added \fr and prerequisite functions. Modified docs.
\fr is a contextually-aware \frac. It will typeset fractions
in the usual \frac style in display mode, and with a slash
otherwise---including when within another \fr. This reduces
the height of a mathematical expression, desirably.

The slashed fraction parenthesizes the arguments if the length
of the argument is over one character. This is imperfect, but
works fine for now.